Added screen brightness PWM control

+++ src/Settings.h
@@ -8,7 +8,7 @@ const uint8_t ShutdownMinutes[] = {0, 5, 10, 15, 20};
 struct SettingsData {
 	bool sound = true;
 	uint8_t sleepTime = 1; //index of value in vector, values : {0, 5, 10, 15, 20}[min]
-	uint32_t screenBrightness = 255;
+	uint8_t screenBrightness = 255;
 	bool tested = false;
 };
